backend/requirements.txt
Django>=5.2
psycopg2-binary>=2.9
djangorestframework>=3.16
django-cors-headers>=4.3 
stripe>=8.0 
gunicorn 
djangorestframework-simplejwt
beautifulsoup4
openai>=1.0.0
anthropic>=0.7.0
langchain>=0.1.0
transformers>=4.35.0
torch>=2.1.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
pillow>=10.0.0
celery>=5.3.0
redis>=5.0.0
django-redis>=5.4.0
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0
django-filter>=23.0
django-guardian>=2.4.0
django-ratelimit>=4.1.0
django-extensions>=3.2.0
django-ckeditor>=6.7.0
django-taggit>=4.0.0
django-simple-history>=3.4.0
django-import-export>=3.3.0
sentry-sdk>=1.40.0
django-health-check>=3.17.0
django-admin-interface>=0.26.0
django-colorfield>=0.11.0
django-mptt>=0.15.0
django-reversion>=5.0.0
requests>=2.31.0
python-dotenv>=1.0.0 